Fixed missing genesis for Mumbain and BorMainnet
It appears the `cmd/cli` package is the new CLI that **must** be used, the `cmd/geth` is definitely broken now. First it contains duplicated flags for Goerli, second problem, `nil pointer` is raised using without further modifications to the codebase, which makes me wonder if there is not other things missing when using `cmd/geth` now.