Changed a bit how we handle OnNewAccount
Reading more and more the original code and the new code, I think we won't be able to actually create a fully 1:1 implementation that mimics old behavior. This is because the way all this is managed drastically changed in recent version of Geth, removing the old `getDeletedStateObject` that was used before. It does not seems easily possible to relicate old behavior in all cases.

I think we will need to let go account creations altogether as they cannot be determistic.